api/upload: Make sure http lib doesn't close the file
The interface of not receiving an io.ReadSeekCloser was exactly
so that we didn't take ownership of losing that reader, but the
caller should do it instead once they're done with the file.

The http lib receives an io.Reader in the constructor as well,
the big problem is that it converts it internally to a closer
and then force that it is the one closing the file.

This wraps the reader with a NopCloser so we are able to handle
closing it ourselves (or rather, our caller).